Numerous tweaks to tpch_prepare shell script.
Ensure that it is possible to avoid WAL-logging whenever that is sane. Add
additional useful instrumentation and input sanitization. Assume that the
ordinary case is that the user wants new data. Make benchmark scale
configurable. Use dbgen in verbose mode for additional information. Add comment
to database indicating when it was created (viewable with '\l+' in psql). Show
elapsed time.